+++ src/OpenSimBindings/ObjectPropertyEdit.hpp
@@ -15,6 +15,11 @@ namespace osc
     class ObjectPropertyEdit final {
     public:
         ObjectPropertyEdit(
+            OpenSim::AbstractProperty const&,
+            std::function<void(OpenSim::AbstractProperty&)>
+        );
+
+        ObjectPropertyEdit(
             OpenSim::Object const&,
             OpenSim::AbstractProperty const&,
             std::function<void(OpenSim::AbstractProperty&)>
@@ -23,6 +28,10 @@ namespace osc
         std::string const& getComponentAbsPath() const;  // empty if it's just a standalone object
         std::string const& getPropertyName() const;
         void apply(OpenSim::AbstractProperty&);
+        std::function<void(OpenSim::AbstractProperty&)> const& getUpdater() const
+        {
+            return m_Updater;
+        }
 
     private:
         std::string m_ComponentAbsPath;